Create OCR_translater.py
Originally `OCR_PATTERNS` was reversed and used lists instead of tuples, and I used a `for: else` loop in `convert()`.  I changed it to how it is though, because it's how it should be.
Everything is only used once, so there was really no need to en-functionate things, but it's nice.  I like it.

@AdaSongS solution is also worth looking into: 
https://exercism.org/tracks/python/exercises/ocr-numbers/solutions/AdaSongS
She used sets of columns to extract out the OCR digits.  This is also what I was considering, but I figured using `wrap()` would probably be easier.  Yes and no.  `wrap()` ended up having some issues and not working, but I was able to write my own wrap without anything extra going on.